Touchable: Pass 2nd argument of `TouchableNativeFeedback.Ripple`.
This gets flagged as required after the upcoming RN v0.61 -> v0.62
upgrade. It's marked as required in the RN v0.61 doc [1], so we
should be passing it already. A value of `false` means the ripple
will stay within the view bounds, which is the status quo:

"""
If property `borderless` evaluates to true the ripple will render
outside of the view bounds (see native actionbar buttons as an
example of that behavior).
"""

[1] https://reactnative.dev/docs/0.61/touchablenativefeedback#ripple