UserList [nfc]: Remove `style` prop.
Investigation was prompted by a Flow error with the `style` prop
that would start appearing at the upcoming RN v0.61 -> v0.62
upgrade.

A good model for getting that prop well-typed would be the
same-named prop of `ZulipButton`, following Greg's adjustments in
8b30fa9d6. Just as we do there, we'd want a subset of `ViewStyle` --
`UserList`'s `style` prop is merged with an internally defined style
object, and the result is passed to `SectionList` as its `style`
prop. `SectionList`'s `style` prop is inherited from `ScrollView`
[1], which is in turn inherited from `View` [2] [3].

However, one key difference from `ZulipButton` is that we haven't
made much use of `UserList`'s style prop; in fact, the only use we
make of it is to provide a redundant `flex: 1` attribute that is
already defined in `UserList`'s internal style variable.

Greg's comment in 8b30fa9d6 suggests that it wouldn't obviously be a
great idea to make more use of `UserList`'s style prop:

```
// The use of specific style properties for callers to micromanage the
// layout of the button is pretty chaotic and fragile, introducing
// implicit dependencies on details of the ZulipButton implementation.
// TODO: Assimilate those usages into a more coherent set of options
//   provided by the ZulipButton interface explicitly.
//
// Until then, [...]
```

So, just remove it.

[1] https://reactnative.dev/docs/sectionlist#props
[2] https://reactnative.dev/docs/scrollview#props
[3] https://reactnative.dev/docs/view#style