apiFetch types: Use `RequestOptions` for `getFetchParams`.
With facebook/react-native#29866, `fetch` and a few of the major
things it uses are totally untyped. But react-native doesn't
suppress `RequestOptions` [1], so we can still use that to help make
sure we're passing around the right kinds of stuff that will
eventually be fed to `fetch`.

Also, it looks like the functionality of `getFetchParams` would
break if a caller included `headers`, so make the type not allow
that.

A type-cast and a $FlowFixMe on it at the top of `getFetchParams`
can fall away naturally.

[1] See it at https://www.saltycrane.com/cheat-sheets/flow-type/latest/