compose [nfc]: Adjust comments on new video-call code.
On the Redux types, this mostly revises to align the comments with
our favorite style guide on doc comments:
  https://dart.dev/guides/language/effective-dart/documentation
in particular by having the type's comment focus on describing a
value of the type, and by focusing on information the reader can't
learn from the type definitions themselves.

In the reducer, the information that was in the doc comment is best
covered on the type definitions; but an implementation comment
pointing to the webapp code this duplicates is helpful.  Also for
this simple local helper function we can drop the annotations on
the argument types.

In ComposeBox, this comment wasn't actionable for the reader, and
isn't worth its cost in taking the reader's attention from how the
existing code works.