deps: Upgrade react-redux to v7.
From the release notes [1]:

"""
This release has undergone extensive performance benchmarking, and
we're confident that it's the fastest version of React-Redux yet!
"""

Cool! It doesn't necessarily mean a dramatic improvement all across
our app (or anywhere); we'd need to do our own testing to establish
that. But this is encouraging.

This version now uses React Hooks in the `connect` implementation,
so they announce that we need to be using React 16.8.4 or later.
That's no problem; we're well past that and have been for a while.

In fact, the corresponding React peer-dep change is announced as the
only breaking public API change.

There's a FlowTyped libdef for this version, so, grab that; there
are no new Flow errors.

[1] https://github.com/reduxjs/react-redux/releases/tag/v7.0.1