ComposeBox: Fix a crash on sending a message.
This was a regression caused by something that was missed in
fd4debbae; we partly but incompletely removed our use of
`getWrappedInstance` with a react-redux API change. From that commit
message:

"""
So, make that change, including the implied removal of
`getWrappedInstance`; that bit of UI works fine on Android and iOS
after that removal.
"""

It looks like "that bit of UI" was something other than sending a
message, and I didn't do a full-project search for the term
`getWrappedInstance`. This should be the last of them now, and I
just tested and I don't get a crash on sending a message.

Fixes: #4270