android build: Allow Gradle to use more memory.
I've been increasingly often seeing the build fail with an error
reading "Java heap space".  This seems to be because the JVM is
run with an upper bound on how much RAM it can use, and we're
hitting the limit.  The bound is already quite a bit (half a GiB)
so that using more seems unnecessary, and it's probably not a
coincidence that the build takes unreasonably long, like >30s...
but having to retry it doesn't make it go any quicker.  So,
increase the heap size permitted.

While here, correct the comment about the default value, and take
the opportunity to have Gradle stop leaving Java heap profiles
lying around on disk when it does fail this way.  I hadn't noticed
that latter config option before, but upon looking closely at this
line, it was staring me in the face.