fetchingReducer: Stop listening for DO_NARROW in the last place we do so.
And remove DO_NARROW.

We greatly reduced the responsibility of the DO_NARROW action in
55361d844, leaving only this site in our reducers.

This last remaining site in `fetchingReducer` resembles a
garbage-collection mechanism, to clear out the `fetching` state so
it can't grow to be too big. But, as Greg says,

"""
Yeah -- I think that isn't really even essential to have at all. The
`fetching` state is a tiny bit of metadata for each narrow, and
whenever we have it we'll have much more data at the same narrow in
`state.narrows`.

The logic there also doesn't totally make sense to me: when we
navigate to a new narrow, we forget all about the fetching state of
other narrows? Basically if that ever results in dropping an entry
where one of the values was `true`, then that seems like a bug to
me. Possibly latent and not visible to the user.
"""