messagesActions: Don't let `doNarrow` cancel an expected navigation (1/2).
As Greg points out in discussion [1], it's not great for `doNarrow`
to silently not navigate when the user expects to navigate.

In the case of an invalid narrow, allow the navigation, but make
sure `ChatScreen` gracefully handles it by showing an error state to
the user. That view naturally looks a lot like the `FetchError` view
added in 353cd3a3e for showing an error in fetching data.

A few tests that have an appropriately narrow focus to confirm the
now-absent behavior of `doNarrow` (checking that the narrow is
valid) are now wrong, so, remove them. A natural question is whether
these tests were giving us more coverage than we already have in
`isNarrowValid`'s tests, and it appears that they're not. In
particular, the two removed tests seem to map neatly onto two of
`isNarrowValid`'s tests, as follows:

(in messageActions-test.js -> in narrowsSelectors-test.js)

'if new narrow stream is not valid, do nothing' ->
  'narrowing to a non-existing stream is invalid'

'if new narrow user is deactivated, do nothing' ->
  'narrowing to a PM with non-existing user is not valid'

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/decouple.20nav.20from.20redux.20%28.23M3804%29/near/1027854