SmartUrlInput: Properly type the `navigation` prop.
As an ordinary, reusable component that wraps something as
multipurpose as the `Input` component, `SmartUrlInput` is meant to
get the `navigation` prop in the normal React way. This is in
contrast to the many screen components that are invoked by passing
them to `create*Navigator` as part of the route config (and in no
other places), and that therefore get their `navigation` prop "for
free".

There's no need to make `SmartUrlInput` too picky about the shape of
the `navigation` prop it receives -- it shouldn't obviously care
that it's the kind of `navigation` prop that you get from a stack
navigator, as opposed to a drawer navigator, etc.

I've found a way to spell that --
`NavigationScreenProp<NavigationState>` -- and it seems to work [1].
In particular, it seems to work nicely with the pattern we'll take
in the next commit of annotating our "screen" components with the
`navigation` prop. So, in this commit, also adopt that pattern for
all the screens that currently use `SmartUrlInput`. Currently,
there's just one, `RealmScreen`.

[1] https://github.com/zulip/zulip-mobile/pull/4273#discussion_r498003062