main screen: Render nothing if no server data.
Since the upgrade to react-navigation v4 (f3b6c1f9e), there's
a crash which seems to be caused by the main screen getting
re-`render`ed while we transition away on logout.  At this
point we've dropped the server data, which much of our main UI
relies on; and there also seems to be a react-navigation bug
where it's already dropped the child navigator (our MainTabs)
from its own data structures, and crashes at the re-render.

Fix that by just rendering nothing in that case.

Fixes: #4275