ios: Drop iOS 10 support; require iOS 11.0+.
As of 2020-09-08, 99.7% of our iOS users are on iOS 11 or later [1].
As Greg says,

"""
With these figures, iOS 11.x joins iOS 10.x in being far into the
range where if we learn things don't work or require effort to keep
working, we'll cheerfully just drop support.
"""

Also, allow graceful degradation for iOS 12. As of Greg's report,
versions <= 12 were 7.3% of iOS users, so graceful degradation at
that level is fine.

The particular feature we'll start using in an upcoming commit,
which requires iOS 11, is "named colors" [2] [3] [4] [5].

[1] https://chat.zulip.org/#narrow/stream/48-mobile/topic/platform.20versions/near/1012197

[2] A pre-official release of Xcode 11 reportedly had a bug with
    this feature; see
    https://developer.apple.com/forums/thread/122661. The discussion
    there suggests that the bug was fixed in the "GM Seed 2" release
    for Xcode 11, and the build number for that release is the same
    as the build number for the official release
    (https://xcodereleases.com/). So we should be fine even if we
    use Xcode 11, which we don't (or I don't, anyway; I'm on 12).

[3] https://developer.apple.com/library/archive/documentation/Xcode/Reference/xcode_ref-Asset_Catalog_Format/Named_Color.html

[4] https://developer.apple.com/documentation/uikit/uicolor/2877380-colornamed?language=objc

[5] https://devblog.xero.com/managing-ui-colours-with-ios-11-asset-catalogs-16500ba48205