ios: Set `rootView.loadingView` instead of `rootView.backgroundColor`.
`rootView` is visible after the launch screen displays and before
our React Native components are mounted and visible. It's therefore
tempting to have `rootView`'s background mimic either the launch
screen or the first thing that appears in the app (or both, if
they're similar), so that it doesn't appear as a "flicker" in
between. In particular, with the upcoming solid brand-colored launch
screen, it's tempting to make `rootView.backgroundColor` also be the
brand color.

The reason this isn't great is that `rootView` isn't only visible at
launch. Every time you rotate the device 90 degrees, it makes an
appearance during the rotation animation [1]. It doesn't obscure
anything, but it's annoying and distracting -- currently, it's
explicitly white, and changing it to the brand color would probably
make it even more jarring.

So, don't set `rootView.backgroundColor` to anything. (The unwanted
white areas in the rotation animation go away as a nice result of
this.) Since we *do* want to show the brand color at launch, to
avoid that flicker mentioned above, we can use RCTRootView's
`loadingView` instead -- it's designed exactly for this.

[1] See, e.g., the GIF at
    https://www.npmjs.com/package/react-native-root-view-background.