i18n: Offer zh_TW instead of zh-Hant.
The existing zh-Hant translation appears to use Taiwanese
vocabulary, so it seems like it was effectively a zh_TW translation
in the first place under another name.  Between the two of them,
the zh_TW translation is the one that's now nearly complete, so
prefer that one.  See chat discussion:
  https://chat.zulip.org/#narrow/stream/58-translation/topic/zh-Hant.20.2F.20zh_TW/near/1044288

There are several ways to name the different zh* translations.
For our choice, see discussion a bit later in that thread:
  https://chat.zulip.org/#narrow/stream/58-translation/topic/zh-Hant.20.2F.20zh_TW/near/1045443
This also aligns with the naming used by both Google and Apple
in the system settings of Android and iOS respectively.

A previous draft of this change, before the react-intl v3 upgrade
in 30e4d19cc, had to add a line to the old `src/i18n/locale.js` to
effectively tell react-intl that for general information about the
locale `zh_TW` it should look to what it has for `zh-Hant`.  That
file no longer exists, because the react-intl function
addLocaleData which it was invoking no longer exists, because
react-intl now relies on the `Intl` API instead:
  https://formatjs.io/docs/react-intl/upgrade-guide-3x/#migrate-to-using-native-intl-apis

Empirically, everything in the app seems to work just as well when
I choose this language as with any other well-translated language,
including `zh-Hans`.

 * The obvious places where we'd be using locale data, namely where
   we show dates and times in the message list... aren't
   internationalized anyway, always showing up in English.

 * Watching the console (where `react-native start` is running, in
   a debug build), I don't see any output that isn't present with
   other languages.