i18n: Sort languages in picker UI by selfname, not English name.
This makes a lot more sense when you're using any language
other than English, and it's what most software seems to do.

The other defensible alternative is to sort by the language tag.
In order to make sense, that requires also displaying the tag
in the UI.  This is what Wikipedia does in its preferences UI
on the web:
  https://en.wikipedia.org/wiki/Special:Preferences
but I'm not sure I've ever seen it anywhere else.