usersReducer: Remove unnecessary check.
The thing that determines whether `realm_users` will be present is
whether we ask for it to be there when we register for events. We do
ask, and we ask properly -- we include `realm_user` in
`fetch_event_types`, as the Zulip API doc [1] tells us to.

If it's missing, and we're still asking for it correctly, it's a
bug. In an ideal world, a check for such a bug would naturally be
done in a validate-and-transform layer -- but the place for that is
at the edge, not the "soft center" [2] of the app. Probably
somewhere like src/api/registerForEvents.js.

[1] https://zulip.com/api/register-queue
[2] https://github.com/zulip/zulip-mobile/blob/master/docs/architecture/crunchy-shell.md