outbox: Properly get user data on sending to cross-realm bots, too.
Also to deactivated users, if that's a possible thing to do (I'm
not sure whether it is.)

This appears to be only a latent bug, with no user-visible
symptoms.  But: if you send a PM with a cross-realm bot as
recipient, and use the debug console to look at the outbox message
in the state, you find that the display_recipient has an entry like
`{email: "welcome-bot@zulip.com", id: -1, full_name: ""}`.

We're apparently relying on the email in the most relevant places
to interpret that value, avoiding visible symptoms... but this would
become a live bug when we transition any of that code to use numeric
IDs, as we've generally been doing.

That bogus data in the ID and name were taken from NULL_USER, in
that fallback over in mapEmailsToUsers.  And that fallback was hit
because we're using the limited `getUsersByEmail` rather than the
comprehensive `getAllUsersByEmail`.  Switch to the latter to fix it.