tools/build-webview: Work in terms of relative paths.
On Windows (at least in Git Bash, which is our recommended way to
set up the dev environment on Windows), the absolute paths end up
looking like `/d/zulip-mobile/…`.  When one of those gets passed to
rsync, it somehow decides to interpret it as a remote location,
where the leading `/d/` means to go to a remote host called `d`.
Naturally, that doesn't work out.

To solve the problem, use relative paths.

This is made slightly tricky by the fact that the two different
arguments we pass to rsync come to us in the form of paths relative
to two different locations: the sources we want to copy are at
particular places relative to the source tree, while the
destination directory comes from the command line and is therefore
relative to the current directory the script was started with.  So
we can't pass them both directly to rsync without doing some kind
of normalization on them first: if we're not at the source tree's
root, then a source path like `src/webview/static` won't be right,
but if we cd to the root then the destination path from the command
line won't be right.  That's the problem we'd previously cut right
through by normalizing everything into absolute paths.

Happily, GNU's coreutils offers not only `readlink` but `realpath`,
which includes a `--relative-to` option.  So we can use that to
turn `dest` relative to the starting directory into a `dest` that's
still relative, but now relative to the source root, just like the
source paths are.

(We already require GNU coreutils; that's the `ensure-coreutils.sh`
script invoked near the top of this one.)

Fixes: #4297