navActions: Stop using @react-navigation/compat.
The compatability layer [1] has eased our migration to React
Navigation v5. Now, remove our use of it here by adapting to the new
APIs.

The ugprade guide [2] describes the changes. There are a lot of
changes to the interface, but it does say the following:

"""
In addition, there have been some changes to the way the navigation
actions work. These changes probably won't affect you if you didn't
do any advanced tasks with these methods.
"""

I don't know where they put the threshold for "advanced", but I
think we should be fine.

They do note (as they do on their "navigating without the navigation
prop" doc [3]) that it's best to use the `navigation` object when
one can:

"""
It's highly recommended to use the methods on the navigation object
instead of using action creators and `dispatch`. It should only be
used for advanced use cases.
"""

But I haven't noticed any major disruptions so far.

[1] https://reactnavigation.org/docs/compatibility
[2] https://reactnavigation.org/docs/upgrading-from-4.x/#action-creators
[3] https://reactnavigation.org/docs/navigating-without-navigation-prop