MainTabs: Stop using @react-navigation/compat.
And switch over to the React Navigation v5 component-based API [1]
for it.

For all screens on `MainTabs`, that means the `navigation` prop gets
a new shape, and the `route` prop is introduced. Also, where we
invoke the screens to attach them to the navigator, the screens'
`navigation` and `route` prop type annotations are checked for
correctness.

Since we newly define the `MainTabs` component ourselves, make it a
Hooks-based component, as that seems to be the way of the future.

[1] https://reactnavigation.org/docs/upgrading-from-4.x/#configuring-the-navigator