ZulipAppContainer: Pass `theme` prop to `AppContainer`.
Both React Navigation v4 and v5 have interfaces for setting the
theme between light and dark, but they're quite different.

One kind of unfortunate difference is that, in v5, using that
interface is effectively required, even though we have a so-far
quite adequate way of managing the theme. React Navigation v5 widely
applies some styling overrides to its provided components (such as
the background of a scene in a bottom-tab navigator) even when we
don't pass a `theme` prop. I'm not sure if it applies them based on
its own arbitrary default, or if it's listening to the system-wide
theme setting. But we're left with a choice between papering over
each of those overridden styles, or taking up their theme interface
and doing something with it.

We take the latter approach because it's not piecemeal. (It would be
tricky to find all the forced styles we'd need to override, and it's
not obvious that we'll always be able to paper over them. One such
fix is to set a `backgroundColor` within `sceneContainerStyle` in
`MainTabs` -- but this wouldn't have been possible even a few months
ago, before @react-navigation/bottom-tabs@5.10.0.)

To smoothen the change toward using the `theme` prop in v5, start
doing so now, while we're still on v4. Theme support is much more
rudimentary in v4 (i.e., the `theme` prop here just takes "dark" or
"light"), but I don't see anything particularly disruptive by
starting to use it now [1], so, might as well.

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/React.20Navigation.20v5.3A.20theme.20colors/near/1061811