nav [nfc]: Remove some soon-to-be-unnecessary comments.
When we're done upgrading to React Navigation 5, including the
removal of @react-navigation/compat (which is completely untyped and
has no existing libdef) comments like these will be unnecessary.
Currently, our navigators don't check that their attached screens
are annotated with the correct shape of the `navigation` prop -- so
we've just been saying they are, in these comments. We'll get those
checks when the upgrade is complete.

We remove them now to reduce the size of the diffs in the upcoming
commits.