deps: Set up react-native-safe-area-context properly.
react-native-safe-area is getting a bit old; it's now been two years
since the last change in the GitHub repo
(miyabi/react-native-safe-area). And, according to NPM,
react-native-safe-area-context has almost 300,000 weekly
downloads [1], to react-native-safe-area's 7,000 [2].

This more popular library gives us a built-in way to consume the
safe-area insets, so that we don't have to go through Redux.
Actually, three ways: a HOC, a React Context consumer, and a React
Hook. The inclusion of the Hook, in particular, signals that this
library is still maintained and forward-looking. See discussion [3]
for one reason we might prefer to stop using Redux to handle the
safe-area insets.

We installed it in f3b6c1f9e to satisfy peer-dependency constraints,
but at that point we hadn't been planning on actually using it. So,
now, make a libdef for it (using Flowgen) and add the
`SafeAreaProvider` to our various wrappers in src/ZulipMobile.js.

[1] https://www.npmjs.com/package/react-native-safe-area-context
[2] https://www.npmjs.com/package/react-native-safe-area
[3] https://github.com/zulip/zulip-mobile/pull/4300#discussion_r524881747