narrow: Add `getNarrowsForMessage`.
We've never had a function for getting all the narrows that a
message appears in. Such a function will be important
algorithmically for avoiding a performance regression with the
EVENT_NEW_MESSAGE action if/when we start storing `state.narrows` as
an `Immutable.Map`; see discussion of this strategy at
https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.60Immutable.2EMap.60.20performance.20for.20.60state.2Enarrows.60/near/1068912

Might as well put it in `narrow.js` and make it reusable, too.