flow: Follow RN's removal of `inexact-spread` lint rule.
Part of the RN v0.62 -> v0.63 changes to the template app [1],
corresponding to facebook/react-native@87370c4c0.

This lint rule was removed in facebook/flow@759970c1b, which was
released in Flow v0.117. So we'd get an error if we left this rule
in when taking Flow v0.122.0 along with the RN upgrade.

But as Greg points out [2]:

"""
the lint had already stopped doing anything in v0.111 at
facebook/flow@0410e0502 -- or at least, it had stopped ever firing
in any situation exercised by Flow's tests. (The later commit that
removed it didn't have to touch the tests, except to remove the lint
from a flowconfig.)
"""

Since we're on v0.113.0 now, we're in that zone where the rule does
nothing. Since the right time to remove obsolete code is ASAP, now
is a fine time to remove the lint from our config. :)

As Greg points out, the removal of this lint rule in 0.117 should
probably have been mentioned in the Flow changelog [3].

[1] https://react-native-community.github.io/upgrade-helper/?from=0.62.2&to=0.63.4
[2] https://github.com/zulip/zulip-mobile/pull/4318#discussion_r535919833
[3] https://github.com/facebook/flow/blob/master/Changelog.md#01170