flipper android: Follow RN's exclusion of okhttp from a Flipper dependency.
Part of the RN v0.62 -> v0.63 changes to the template app [1],
corresponding to facebook/react-native@a642c3f26.

This seems appropriate to do right away; a comment [2] on the
corresponding RN issue links (as "the culprit") to a line in v0.33.1
of Flipper [3]. We're on v0.39.0, so it seems like it's time for us
to take this fix.

The consequence of not making this change is reported to be an
inability to develop/debug with Android API < 21. There are
reportedly no effects in production.

[1] https://react-native-community.github.io/upgrade-helper/?from=0.62.2&to=0.63.4
[2] https://github.com/facebook/react-native/issues/28481#issuecomment-645546195
[3] https://github.com/facebook/flipper/blob/v0.33.1/build.gradle#L89