zulipVersion: Expose `.elements` instead of `.loggingArray`.
We'll call `.elements` to get what we need to make tags to send to
Sentry.

As Greg points out [1],

"""
I think it probably makes most sense for the code that's preparing
data for Sentry to be responsible for what exact set of tags we send
to Sentry and how they're assembled.
"""

`.elements` gives us all the data we've parsed from the raw string,
in a nice shape -- and it doesn't try to second-guess what we might
want to send to Sentry, e.g., by replacing a missing minor or patch
version with a zero.

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.23M3745.20Server.20version.20to.20Sentry/near/1059049