zulipVersion [nfc]: Make two methods pure functions.
Also, call `._getElements` just once, instead of twice.

These methods' implementations don't need to be aware of a
ZulipVersion instance's state, or anything. They just take some
input, do some computations on it, and return an output. So it makes
sense for them to be written as pure functions.

They're closely tied to the details of `ZulipVersion`'s
implementation, so it makes sense for them to be static methods
rather than, e.g., module-level, unexported functions [1].

[1] https://github.com/zulip/zulip-mobile/pull/4326#discussion_r535894526