types: Give a real type for display_recipient, on Message and Outbox.
Now the type-checker can confirm that we explicitly handle the two
possibilities here.  The previous commit set us up so that this passes
Flow, mostly by centralizing accesses to display_recipient through a
pair of helpers that explicitly confirm the type is the one we expect.

There are still a number of situations where if we're sloppy about
this the type-checker won't catch it: in particular if we pass the
display_recipient to `===` or `JSON.stringify`, which accept anything,
or get its `.length` which exists on both strings and arrays.  We'll
clean those up too in upcoming commits.