recipient [nfc]: Simplify interface at normalizeRecipients.
That union was indeed confusing... and it turned out it was serving as
basically a roundabout $FlowFixMe for one of this function's callers!
Fix that type issue directly, and simplify.  All callers do indeed
always pass an actual array -- and the behavior would be broken or
meaningless anyway if they didn't.