recipient: Collect all copies of "PM key recipients" filter logic.
This exists several times over because we do it on emails, or on
user IDs, or on user IDs found at a different property name.  It'd
be possible to deduplicate it down to just one copy with a bit more
abstraction... but it's short enough that the trade wouldn't be
worth it, and we haven't.  Still, we can at least make all the
copies look completely parallel and live right next to each other.

This is not quite NFC because in the ...ByEmail case, when there's
just one recipient in the original list (which, for the
display_recipients of a message, happens only for a self-PM),
passing it through normalizeRecipients causes the same whitespace
fudges to be applied as we already do the rest of the time.  Those
fudges probably are never doing anything, and definitely should get
removed from there one way or another; a few commits ago we added
some logging on them as a step toward doing so.