logging [nfc]: Stop importing store, by injecting the dependency.
Before this change, we have our logging code importing the module with
our Redux store, in order to get the version of the Zulip server we're
currently talking to so it can include that as context in the error
report.  The logging code is naturally low in the import graph
(imported by lots of other modules) and the app's Redux store is
naturally high in the import graph (importing lots of other
modules)... so this tends to lead to import cycles and to large blobs
of import cycles.

At present, this import only participates in a couple of small cycles;
the whole blob looks like this:

    $ f=src/utils/logging.js; npx flow cycle --strip-root "${f}" \
        | perl -lne 'print $1 if (/"(.*?)"/)' | sort -u
    src/boot/store.js
    src/boot/ZulipAsyncStorage.js
    src/redux-persist-migrate/index.js
    src/utils/logging.js

But the immediate motivation for this change is that we'd like to add
some logging in src/utils/recipient.js.  That's another module
naturally low in the import graph, so there are a number of modules
importing it that are in turn imported (transitively) by store.js,
making the cycle-blob (or "strongly-connected component": the set of
modules which have import chains both to and from a given module and
therefore each other) look like this:

    src/boot/reducers.js
    src/boot/store.js
    src/boot/ZulipAsyncStorage.js
    src/caughtup/caughtUpReducer.js
    src/chat/fetchingReducer.js
    src/chat/narrowsReducer.js
    src/redux-persist-migrate/index.js
    src/typing/typingReducer.js
    src/unread/unreadHuddlesReducer.js
    src/unread/unreadPmsReducer.js
    src/utils/logging.js
    src/utils/narrow.js
    src/utils/recipient.js

And as a result it produces confusing errors in Flow, and failures in
Jest where whole imported modules are just `undefined` because they
haven't been initialized yet.

Happily, by severing the one link from logging.js to store.js, we
completely dissolve this cycle-blob!  With this change, and even with
that `recipient` -> `logging` import added, neither src/boot/store.js
nor src/utils/logging.js nor any of the other modules they were in
import cycles with is part of any import cycle.

To accomplish that import-severing, we use dependency injection:
the logging module no longer knows where to find the store, and
instead that information gets handed to it.  We have the store
module do this at its own initialization time; if we're logging
something before the store even exists, it wouldn't have had any
meaningful information anyway.