narrow [nfc]: Add constructor pm1to1NarrowFromUser.
This is just like the old pmNarrowFromEmail, but takes a whole user
object.  Using this function guarantees that we *have* a user object
on hand -- and therefore not just an email but a user ID.  Once we're
using this and other such constructors exclusively, that will enable
us to start identifying PM narrows by user IDs, rather than by emails.