recipient [nfc]: Sort (by user ID) in pmKeyRecipientsFromMessage.
This also starts us sorting by ID in pmUiRecipientsFromMessage.  But
pmKeyRecipientsFromMessage is more interesting, because there it'll
shortly make sense to make that part of its interface.  In practice,
for both functions it's already sorted because that's how a group PM's
display_recipient comes from the server.

The similar change in filterRecipientsAsUserIds is there for
parallelism, but has no effect because its caller already sorts the
same way anyway.

On filterRecipientsByEmail we don't have the user IDs, so we can't do
our standard sorting.  This is OK because its only caller does its own
sorting anyway.