recipient [nfc]: One more validating helper; convert the last PM callsite!
We're now down to zero callsites (in non-test code) that call
pmNarrowFromEmails directly!  All callers that construct a PM narrow
now use one of our new constructors that require types containing
user IDs as well as emails, and constructible only by helpers that
guarantee they properly filter and sort the list.

This change is NFC because this callsite was already sorting the list,
and the users in it are chosen from a UI which doesn't include the
self user in the first place.