redux: Store `state.messages` as an `Immutable.Map`.
An instance of #3949 and #3950, like 17bd75231.

One important difference from 17bd75231 is that the `state.messages`
map has numeric keys. We've handled the replace/revive logic for
that, in the previous commit. But care must be taken whenever we use
an `Immutable.Map(...)` call when we want a map with numeric keys,
If you pass an object-as-map to `Immutable.Map`, it's impossible for
the resulting `Immutable.Map`'s keys to be numbers, because the
object-as-map's keys are necessarily strings, and that's up to
JavaScript. The solution is to pass an array of key-value pairs, but
unfortunately, Flow won't catch it if you forget to do this.