notif: Look up user promptly in server data, for opening 1:1 notif.
We already do this for opening a notification for a group PM; now do
the same for 1:1 PMs too.  If we manage to have a notification where
we don't know of a sender with that email -- which can happen if
e.g. the sender has changed it since the notification was originally
sent -- then this will cause us not to try to navigate to the
conversation at all, which is an improvement over anything the
previous behavior might have been: navigating to a conversation
where we show something broken-looking, or crash.

The motivation for this is so that we pass the Narrow constructor a
user ID as well as an email, so that we can start storing user IDs in
Narrow objects and start using them in all the code that consumes them.
This is the last spot in our codebase where we're not doing so.

Really we should use the sender ID from the server.  That started in
Zulip Server 1.8.0, according to the handy comments in FcmMessage.kt.
We leave that be for now, though -- this email use is easy to find
with a grep for `email`.