narrow: Use user IDs in isMessageInNarrow, instead of emails.
The way this change might affect behavior is if we receive a message
where the email for some user doesn't agree (e.g. because the user
changed their email) with one we've seen before, in particular one
we've previously used to identify that narrow in `state.narrows`.
The old code would fail to make the connection, and so wouldn't put
the new message in the existing narrow, which could make it fail to
show up in the message list for that conversation.

After this change, we'll correctly place the new message in the
narrow.