user: Add "...ById" variant for UserAvatarWithPresence.
We'll temporarily have two variants of this component -- an
implementation in common but with slightly different interfaces -- as a
way to smoothly migrate from callers using one interface to the other.

The new interface keeps callers from having to know or care about
emails, which we're moving away from.  It also encapsulates the need
to find the user's avatar URL; the caller only has to be concerned
with identifying which user it wants represented, by user ID.  This
improves react-redux hygiene: it means that when that user's avatar
changes, we no longer have to rerender the calling component, which
fundamentally shouldn't care about the specific avatar URL.