user: Take simply user IDs at GroupDetailsScreen, not whole users.
This fixes a latent bug where if some detail of one of the users
changes while we're showing this screen (or have it in the nav stack),
the nav prop would continue to have the old version of the user data.

(This isn't a live bug because currently the only detail we take
updates for is avatar_url, and UserItem doesn't consult the avatar_url
at all.  It would have been a live bug before the recent 4116f83af.)