UserPickerCard [nfc]: Use setState correctly; also simplify.
Calling `setState` with values based on data we've previously read
from the state is buggy, and explicitly forbidden by the React API:
  https://reactjs.org/docs/state-and-lifecycle.html#state-updates-may-be-asynchronous
The reason is that the actual change to the state may be applied
asynchronously, and so several updates may race each other.

Instead, the proper thing to do is to call `setState` with a callback
that can take the state and compute the update.  Do that.

While here, we can simplify the code a bit by inlining trivial array
manipulations.