messagesReducer: Throw if `message.flags` missing on `eventNewMessage`.
We don't expect this to happen, and a similar check has been done
for the same Redux action in narrowsReducer for some time now. But
we haven't yet made it quite clear to Flow that `flags` will
normally be present here, and we'll want to do that for an upcoming
change.