user types: Use makeUserId in test code.
Where we make up a user ID in test code, mark it in the type system
as being a user ID and not just any old number.

We'll soon adjust the types of our many functions that expect data
containing a user ID, so that they require this marking.  This change
prepares for that.