isInternalLink [nfc]: Simplify logic a bit more.
Each of these three cases can only ever use one of the two branches
of the definition of `restUrl`.

If the URL starts with `/`, then it cannot start with realmStr;
the latter is an absolute URL, and must start with a scheme.

If the URL starts with realmStr, then... it starts with realmStr.

If we make it past the second case and to the third, then the URL
didn't start with realmStr.