isInternalLink: Tighten patterns; write jsdoc.
The old patterns would match a fragment like `#narrowly`, if for some
reason that ever appeared -- which, with the path-like structure that
Zulip gives these fragments, conceptually should be something completely
unrelated.  It's easy to be specific here, so let's do so for clarity.

And with that, I think we've worked out what this function is supposed
to do!  We can say so, and reduce the TODO comment.