store [nfc]: Inline remotedev-serialize (1/x).
It turns out that remotedev-serialize is doing some tasks that we
don't need it to do. See discussion for more background [1].

The changes we'd like to make to fix that will reduce the amount of
code used to implement our replace/revive logic, and it'll make
sense to just fold remotedev-serialize into our codebase instead of
depending on it as a separate module.

One particular thing we want to do is cut `jsan`, something used by
remotedev-serialize, out of the picture. One main feature `jsan`
announces is that it can handle serializing an object with circular
references -- but we clearly don't need this, since our data is
naturally free of them. And as it checks for circular references, it
makes a deep copy of the whole data structure. We also don't need
any of the other fancy features `jsan` advertises, like persisting
functions. [2] So we can potentially gain a good performance boost
by just cutting `jsan` out of the picture. We'll do that after all
these "Inline remotedev-serialize (n/x)" commits.

Another thing (and I opened this as zulip/remotedev-serialize#2
before we decided to inline remotedev-serialize), we've observed [3]
that we can save maybe >20% of the time we periodically spend
persisting the state by removing various lines in
remotedev-serialize itself (not `jsan`) that handle Immutable.Map
data structures that we don't currently use.

Start by cutting out the `Serialize.immutable` factory and just
writing out `stringify` and `parse` (previously obtained by calling
`Serialize.immutable`) ourselves, based on `Serialize.immutable`'s
implementation and the inputs we'd been giving it.

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.60remotedev-serialize.60.20and.20.60jsan.60/near/1082744

[2] It turns out that the circular reference-handling code does
    helpfully recurse through and apply our custom replacer function
    on values before they've been transformed by their `toJSON`
    methods. We'll handle that wrinkle when we take `jsan` out, a
    few commits from now.

[3] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/user_avatar_url_field_optional/near/1082087